# packages/shared/src/generated/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kanuni API",
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Get Liveness",
        "description": "Report whether the process is alive.\n\nReturns:\n    A body indicating the process is running. Never checks downstream\n    dependencies \u2014 that is the job of /readyz.",
        "operationId": "get_liveness_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LivenessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Get Readiness",
        "description": "Report whether the service is ready to accept traffic.\n\nChecks database connectivity. Embedding-model readiness will be added\nonce retrieval is implemented (Phase 2).\n\nArgs:\n    response: The outgoing response, whose status code is set to 503\n        when a dependency is not ready.\n    settings: Application settings, providing the database DSN.\n\nReturns:\n    A body listing the result of each dependency check.",
        "operationId": "get_readiness_readyz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "description": "List documents in the registry, optionally filtered.\n\nArgs:\n    connection: Database connection (injected).\n    _api_key: The authenticated caller's key (requires `query` scope).\n    status: If given, only documents with this status are returned.\n    doc_type: If given, only documents of this type are returned.\n    limit: Maximum number of results.\n    offset: Number of results to skip, for pagination.\n\nReturns:\n    Matching documents, most recently ingested first.",
        "operationId": "list_documents_v1_documents_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "doc_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Doc Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummary"
                  },
                  "title": "Response List Documents V1 Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "description": "Fetch a single document by id.\n\nArgs:\n    document_id: The document's id.\n    connection: Database connection (injected).\n    _api_key: The authenticated caller's key (requires `query` scope).\n\nReturns:\n    The document.\n\nRaises:\n    DocumentNotFoundError: If no document with that id exists.",
        "operationId": "get_document_v1_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/documents": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Upload Document",
        "description": "Upload a document for ingestion.\n\nValidates and dedupes by SHA-256; a re-upload of an already-ingested\nfile is reported as a no-op rather than a duplicate document (\u00a77 stage 1).\n\nArgs:\n    connection: Database connection (injected).\n    _api_key: The authenticated caller's key (requires `ingest:admin` scope).\n    settings: Application settings (for max upload size).\n    storage: Document storage backend (injected).\n    file: The uploaded PDF.\n    source_id: The slug matching an entry in `sources.yaml`.\n    title: The document's title.\n    doc_type: The kind of regulatory instrument.\n    language: The document's language code (e.g. `\"en\"`, `\"sw\"`).\n    reference_number: The document's own reference number, if already known.\n    issued_date: The document's issue date, if already known.\n\nReturns:\n    201 with `{\"status\": \"created\", \"document_id\": ...}` for a new\n    document, or 200 with `{\"status\": \"skipped\", \"document_id\": ...}`\n    for a re-upload of an already-ingested file.",
        "operationId": "upload_document_v1_admin_documents_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_v1_admin_documents_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/ingestion-jobs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Failed Ingestion Jobs",
        "description": "List documents whose ingestion pipeline is currently failed.\n\nArgs:\n    connection: Database connection (injected).\n    _api_key: The authenticated caller's key (requires `ingest:admin` scope).\n\nReturns:\n    One entry per failed document, most recent failure first.",
        "operationId": "list_failed_ingestion_jobs_v1_admin_ingestion_jobs_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngestionJobSummary"
                  },
                  "title": "Response List Failed Ingestion Jobs V1 Admin Ingestion Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/ingestion-jobs/{document_id}/retry": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Retry Ingestion Job",
        "description": "Reset a failed document so the worker reprocesses it from the start.\n\nArgs:\n    document_id: The document to retry.\n    connection: Database connection (injected).\n    _api_key: The authenticated caller's key (requires `ingest:admin` scope).\n\nReturns:\n    A status acknowledgement.",
        "operationId": "retry_ingestion_job_v1_admin_ingestion_jobs__document_id__retry_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Retry Ingestion Job V1 Admin Ingestion Jobs  Document Id  Retry Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/retrieve": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Debug Retrieve",
        "description": "Run the hybrid retrieval pipeline and return scored chunks, for debugging (\u00a714 Phase 2).\n\nArgs:\n    connection: Database connection (injected).\n    _api_key: The authenticated caller's key (requires `ingest:admin` scope).\n    settings: Application settings (retrieval thresholds).\n    embedding_provider: Provider used to embed the question.\n    reranker_provider: Provider used for cross-encoder reranking.\n    question: The question to retrieve chunks for.\n    include_historical: If True, include superseded/repealed documents.\n\nReturns:\n    Up to `settings.rerank_top_k` chunks with every stage's score attached.",
        "operationId": "debug_retrieve_v1_admin_retrieve_get",
        "parameters": [
          {
            "name": "question",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question"
            }
          },
          {
            "name": "include_historical",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Historical"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoredChunk"
                  },
                  "title": "Response Debug Retrieve V1 Admin Retrieve Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/query": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Post Query",
        "description": "Answer a question, streamed over SSE: `token` events, then one `done` event.\n\nArgs:\n    request: The question and optional retrieval overrides.\n    connection: Database connection (injected).\n    api_key: The authenticated caller's key (requires `query` scope).\n    settings: Application settings.\n    embedding_provider: Provider used to embed the question.\n    reranker_provider: Provider used for cross-encoder reranking.\n    llm_provider: Provider used for answer generation.\n\nReturns:\n    An SSE stream. See `QueryResultMetadata` for the final `done`\n    event's payload shape.",
        "operationId": "post_query_v1_query_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_document_v1_admin_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "doc_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "reference_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Number"
          },
          "issued_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued Date"
          }
        },
        "type": "object",
        "required": [
          "file",
          "source_id",
          "title",
          "doc_type"
        ],
        "title": "Body_upload_document_v1_admin_documents_post"
      },
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "in_force",
          "superseded",
          "repealed",
          "unknown"
        ],
        "title": "DocumentStatus",
        "description": "Whether a document is currently authoritative, per PROJECT_SPEC.md \u00a76."
      },
      "DocumentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "doc_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "jurisdiction": {
            "type": "string",
            "title": "Jurisdiction"
          },
          "issuing_body": {
            "type": "string",
            "title": "Issuing Body"
          },
          "reference_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Number"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "issued_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued Date"
          },
          "effective_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "pipeline_status": {
            "$ref": "#/components/schemas/PipelineStage"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "title",
          "doc_type",
          "jurisdiction",
          "issuing_body",
          "reference_number",
          "language",
          "issued_date",
          "effective_date",
          "status",
          "pipeline_status"
        ],
        "title": "DocumentSummary",
        "description": "A `documents` row, as exposed by the registry-browsing endpoints."
      },
      "DocumentType": {
        "type": "string",
        "enum": [
          "circular",
          "act",
          "regulation",
          "notice",
          "guideline"
        ],
        "title": "DocumentType",
        "description": "The kind of regulatory instrument a document is, per PROJECT_SPEC.md \u00a76."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestionJobSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "stage": {
            "$ref": "#/components/schemas/PipelineStage"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "error_details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Details"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "stage",
          "attempt_count",
          "error_details"
        ],
        "title": "IngestionJobSummary",
        "description": "An `ingestion_jobs` row, as exposed by the admin failed-jobs endpoint."
      },
      "LivenessResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "LivenessResponse",
        "description": "Response body for the liveness check."
      },
      "PipelineStage": {
        "type": "string",
        "enum": [
          "fetched",
          "extracted",
          "chunked",
          "embedded",
          "indexed",
          "failed"
        ],
        "title": "PipelineStage",
        "description": "The ingestion pipeline stage a document has most recently completed."
      },
      "QueryRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "include_historical": {
            "type": "boolean",
            "title": "Include Historical",
            "default": false
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "description": "The request body for `POST /v1/query`."
      },
      "ReadinessResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "checks": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Checks"
          }
        },
        "type": "object",
        "required": [
          "status",
          "checks"
        ],
        "title": "ReadinessResponse",
        "description": "Response body for the readiness check, detailing each dependency probed."
      },
      "ScoredChunk": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "section_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Ref"
          },
          "page_start": {
            "type": "integer",
            "title": "Page Start"
          },
          "page_end": {
            "type": "integer",
            "title": "Page End"
          },
          "dense_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dense Score"
          },
          "sparse_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sparse Score"
          },
          "fused_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fused Score"
          },
          "rerank_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Score"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "content",
          "section_ref",
          "page_start",
          "page_end"
        ],
        "title": "ScoredChunk",
        "description": "A chunk candidate carrying whichever retrieval-stage scores have been computed so far."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}